scripts/requirements.txt
python-dotenv
requests
pandas
sklearn
google-cloud-vision